# docs/upgrade/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReVector Core Engine",
    "description": "True editable geometry. No raster wrappers. Deploy behind JerseyOS authentication.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/revector/projects": {
      "post": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Create",
        "operationId": "create_api_revector_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Get",
        "operationId": "get_api_revector_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Delete",
        "operationId": "delete_api_revector_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/status": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Status",
        "operationId": "status_api_revector_projects__project_id__status_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/parts": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Parts",
        "operationId": "parts_api_revector_projects__project_id__parts_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/validation": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Validation",
        "operationId": "validation_api_revector_projects__project_id__validation_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/exports": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Exports",
        "operationId": "exports_api_revector_projects__project_id__exports_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/settings": {
      "put": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Settings",
        "operationId": "settings_api_revector_projects__project_id__settings_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/segments/{part_id}/confirm": {
      "post": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_revector_segments__part_id__confirm_post",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/segments/{part_id}/update": {
      "post": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Update",
        "operationId": "update_api_revector_segments__part_id__update_post",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/parts/actions": {
      "post": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Part Action",
        "operationId": "part_action_api_revector_parts_actions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/segments/{part_id}/vector-edit": {
      "post": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Vector Edit",
        "operationId": "vector_edit_api_revector_segments__part_id__vector_edit_post",
        "parameters": [
          {
            "name": "part_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Part Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VectorEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/revector/projects/{project_id}/artifacts/{artifact_path}": {
      "get": {
        "tags": [
          "projects and manual corrections"
        ],
        "summary": "Artifact",
        "operationId": "artifact_api_revector_projects__project_id__artifacts__artifact_path__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProject": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name",
            "default": "Untitled"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "settings": {
            "$ref": "#/components/schemas/ProcessingSettings"
          },
          "production_specifications": {
            "items": {
              "$ref": "#/components/schemas/PartRequirement"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Production Specifications"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CreateProject"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartAction": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "action": {
            "type": "string",
            "enum": [
              "add",
              "remove",
              "merge",
              "split"
            ],
            "title": "Action"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          },
          "polygon": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array",
                "maxItems": 10000,
                "minItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon"
          },
          "polygons": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "prefixItems": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "number"
                      }
                    ],
                    "type": "array",
                    "maxItems": 2,
                    "minItems": 2
                  },
                  "type": "array"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygons"
          },
          "part_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Ids"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "title": "Name",
            "default": "Manual part"
          },
          "type": {
            "type": "string",
            "enum": [
              "front_body",
              "back_body",
              "left_sleeve",
              "right_sleeve",
              "left_shoulder",
              "right_shoulder",
              "front_collar",
              "back_collar",
              "left_cuff",
              "right_cuff",
              "trim",
              "top_trim",
              "bottom_trim",
              "other_part",
              "unknown"
            ],
            "title": "Type",
            "default": "unknown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "action"
        ],
        "title": "PartAction"
      },
...
      "PartUpdate": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "polygon": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array",
                "maxItems": 10000,
                "minItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "front_body",
                  "back_body",
                  "left_sleeve",
                  "right_sleeve",
                  "left_shoulder",
                  "right_shoulder",
                  "front_collar",
                  "back_collar",
                  "left_cuff",
                  "right_cuff",
                  "trim",
                  "top_trim",
                  "bottom_trim",
                  "other_part",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "locked": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked"
          },
          "confirmed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed"
          },
          "physical_width_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Width Mm"
          },
          "physical_height_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Height Mm"
          },
          "bleed_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bleed Mm"
          },
          "safe_zone_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Zone Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "PartUpdate"
      },
...
      "SettingsUpdate": {
        "properties": {
          "ai_workflow": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Workflow"
          },
          "mockup_width": {
            "anyOf": [
              {
                "type": "integer",
                "multipleOf": 16.0,
                "maximum": 1920.0,
                "minimum": 1024.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mockup Width"
          },
          "mockup_height": {
            "anyOf": [
              {
                "type": "integer",
                "multipleOf": 16.0,
                "maximum": 1920.0,
                "minimum": 768.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mockup Height"
          },
          "preset": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "FAST",
                  "BALANCED",
                  "PRECISION",
                  "ULTRA"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset"
          },
          "vector_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "precision",
                  "color",
                  "mono",
                  "reconstruction"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Mode"
          },
          "max_colors": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 64.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Colors"
          },
          "delta_e": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta E"
          },
          "min_region_area": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Region Area"
          },
          "segment_min_area_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 0.9,
                "minimum": 0.0001
              },
              {
                "type": "null"
              }
            ],
            "title": "Segment Min Area Ratio"
          },
          "noise_reduction": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Noise Reduction"
          },
          "preserve_original_colors": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preserve Original Colors"
          },
          "ocr": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr"
          },
          "text_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "outlined",
                  "editable"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Mode"
          },
          "gradients": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gradients"
          },
          "allow_contour_fallback": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allow Contour Fallback"
          },
          "max_trace_dimension": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12000.0,
                "minimum": 64.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Trace Dimension"
          },
          "export_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "true_vector",
                  "hybrid"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Export Mode"
          },
          "known_width_mm": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Known Width Mm"
          },
          "bleed_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bleed Mm"
          },
          "safe_zone_mm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Zone Mm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsUpdate",
        "description": "Partial settings are validated against the stored project after merging."
      },
...
      "StageRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "part_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "StageRequest"
      },
...
    }
  }
}